feat: add delete
