+++ src/utils/constants/mediaList.ts
@@ -10,54 +10,66 @@ export const mediaList: mediaListType[] = [
     id: 0,
     category: '카테고리1',
     title: '영상 제목입니다. 두 줄까지 허용됩니다.',
-    description: '영상 메모 영역입니다. 2줄까지 하용됩니다.',
+    description: '영상 메모 영역입니다. 2줄까지 허용됩니다.',
   },
   {
     id: 1,
     category: '카테고리2',
     title: '영상 제목입니다. 두 줄까지 허용됩니다.',
-    description: '영상 메모 영역입니다. 2줄까지 하용됩니다.',
+    description: '영상 메모 영역입니다. 2줄까지 허용됩니다.',
   },
   {
     id: 2,
     category: '카테고리3',
     title: '영상 제목입니다. 두 줄까지 허용됩니다. ',
-    description: '영상 메모 영역입니다. 2줄까지 하용됩니다.',
+    description: '영상 메모 영역입니다. 2줄까지 허용됩니다.',
   },
   {
     id: 3,
     category: '카테고리1',
     title: '영상 제목입니다. 두 줄까지 허용됩니다. ',
-    description: '영상 메모 영역입니다. 2줄까지 하용됩니다.',
+    description: '영상 메모 영역입니다. 2줄까지 허용됩니다.',
   },
   {
     id: 4,
     category: '카테고리2',
     title: '영상 제목입니다. 두 줄까지 허용됩니다. .',
-    description: '영상 메모 영역입니다. 2줄까지 하용됩니다.',
+    description: '영상 메모 영역입니다. 2줄까지 허용됩니다.',
   },
   {
     id: 5,
     category: '카테고리3',
     title: '영상 제목입니다. 두 줄까지 허용됩니다. ',
-    description: '영상 메모 영역입니다. 2줄까지 하용됩니다.',
+    description: '영상 메모 영역입니다. 2줄까지 허용됩니다.',
   },
   {
     id: 6,
     category: '카테고리1',
     title: '영상 제목입니다. 두 줄까지 허용됩니다. ',
-    description: '영상 메모 영역입니다. 2줄까지 하용됩니다.',
+    description: '영상 메모 영역입니다. 2줄까지 허용됩니다.',
   },
   {
     id: 7,
     category: '카테고리2',
     title: '영상 제목입니다. 두 줄까지 허용됩니다. .',
-    description: '영상 메모 영역입니다. 2줄까지 하용됩니다.',
+    description: '영상 메모 영역입니다. 2줄까지 허용됩니다.',
   },
   {
     id: 8,
     category: '카테고리3',
     title: '영상 제목입니다. 두 줄까지 허용됩니다. ',
-    description: '영상 메모 영역입니다. 2줄까지 하용됩니다.',
+    description: '영상 메모 영역입니다. 2줄까지 허용됩니다.',
+  },
+  {
+    id: 9,
+    category: '카테고리3',
+    title: '영상 제목입니다. 두 줄까지 허용됩니다. ',
+    description: '영상 메모 영역입니다. 2줄까지 허용됩니다.',
+  },
+  {
+    id: 10,
+    category: '카테고리3',
+    title: '영상 제목입니다. 두 줄까지 허용됩니다. ',
+    description: '영상 메모 영역입니다. 2줄까지 허용됩니다.',
   },
 ];
